Rename Card to CardValidator
