+++ CardValidator.rb
